rewrote learning objective for trees
There are two components of the objective, now: (1) the best-effort
component, and (2) the topological component. The topological component
counts the number of split and merges, the best-effort the number of
successfully segmented best-effort objects. Both components are combined
into a single objective by scaling the topological error in the interval
[0,1) and adding it to the integral best-effort error. This way,
solutions that recover more best-effort regions are prefered, even if
the number of split and merges is higher.